requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-appconf==1.0.2
django-environ==0.4.5
django-filter==2.1.0
django-model-utils==3.1.2
djangorestframework==3.9.4
idna==2.8
Markdown==3.1.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
mysqlclient==1.4.2.post1